+++ config/config.exs
@@ -66,6 +66,7 @@ config :pleroma, Pleroma.Upload,
   proxy_remote: false,
   filename_display_max_length: 30,
   base_url: nil,
+  all_base_urls: nil,
   allowed_mime_types: ["image", "audio", "video"]
 
 config :pleroma, Pleroma.Uploaders.Local, uploads: "uploads"